+++ graphql/resolvers/index.js
@@ -1,164 +1,11 @@
-const bcrypt = require('bcryptjs');
+const authResolver = require('./auth');
+const eventsResolver = require('./events');
+const bookingResolver = require('./booking');
 
-const Event = require('../../models/events');
-const User = require('../../models/user');
-const Booking = require('../../models/bookings');
-
-const events = async eventIds => {
-  try {
-    const events = await Event.find({ _id: { $in: eventIds } });
-    events.map(event => {
-      return {
-        ...event._doc,
-        _id: event.id,
-        date: new Date(event._doc.date).toISOString(),
-        creator: user.bind(this, event.creator)
-      };
-    });
-    return events;
-  } catch (err) {
-    throw err;
-  }
+const rootResolver = {
+  ...authResolver,
+  ...eventsResolver,
+  ...bookingResolver
 };
 
-const singleEvent = async eventId => {
-  try {
-    const event = await Event.findById(eventId);
-    return {
-      ...event._doc,
-      _id: event.id,
-      creator: user.bind(this, event.creator)
-    };
-  } catch (err) {
-    throw err;
-  }
-};
-
-const user = async userId => {
-  try {
-    const user = await User.findById(userId);
-    return {
-      ...user._doc,
-      _id: user.id,
-      createdEvents: events.bind(this, user._doc.createdEvents)
-    };
-  } catch (err) {
-    throw err;
-  }
-};
-
-module.exports = {
-  events: async () => {
-    try {
-      const events = await Event.find();
-      return events.map(event => {
-        return {
-          ...event._doc,
-          _id: event.id,
-          date: new Date(event._doc.date).toISOString(),
-          creator: user.bind(this, event._doc.creator)
-        };
-      });
-    } catch (err) {
-      throw err;
-    }
-  },
-  bookings: async () => {
-    try {
-      const bookings = await Booking.find();
-      return bookings.map(booking => {
-        return {
-          ...booking._doc,
-          _id: booking.id,
-          user: user.bind(this, booking._doc.user),
-          event: singleEvent.bind(this, booking._doc.event),
-          createdAt: new Date(booking._doc.createdAt).toISOString(),
-          updatedAt: new Date(booking._doc.updatedAt).toISOString()
-        };
-      });
-    } catch (err) {
-      throw err;
-    }
-  },
-  createEvent: async args => {
-    const event = new Event({
-      title: args.eventInput.title,
-      description: args.eventInput.description,
-      price: +args.eventInput.price,
-      date: new Date(args.eventInput.date),
-      creator: '6462085a6e9f7493f1c2b53f'
-    });
-    let createdEvent;
-    try {
-      const result = await event.save();
-      createdEvent = {
-        ...result._doc,
-        _id: result._doc._id.toString(),
-        date: new Date(event._doc.date).toISOString(),
-        creator: user.bind(this, result._doc.creator) // creator contains id, email, password, createdEvents
-      };
-      const creator = await User.findById('6462085a6e9f7493f1c2b53f');
-
-      if (!creator) {
-        throw new Error('User not found.');
-      }
-      creator.createdEvents.push(event);
-      await creator.save(); // Update the user db
-
-      return createdEvent;
-    } catch (err) {
-      console.log(err);
-      throw err;
-    }
-  },
-  createUser: async args => {
-    try {
-      const existingUser = await User.findOne({ email: args.userInput.email });
-      if (existingUser) {
-        throw new Error('User exists already.');
-      }
-      const hashedPassword = await bcrypt.hash(args.userInput.password, 12);
-
-      const user = new User({
-        email: args.userInput.email,
-        password: hashedPassword
-      });
-
-      const result = await user.save();
-
-      return { ...result._doc, password: null, _id: result.id };
-    } catch (err) {
-      throw err;
-    }
-  },
-  bookEvent: async args => {
-    const fetchedEvent = await Event.findOne({ _id: args.eventId });
-    const booking = new Booking({
-      user: '646208d06e9f7493f1c2b542',
-      event: fetchedEvent
-    });
-    const result = await booking.save();
-    return {
-      ...result._doc,
-      _id: result.id,
-      user: user.bind(this, booking._doc.user),
-      event: singleEvent.bind(this, booking._doc.event),
-      createdAt: new Date(result._doc.createdAt).toISOString(),
-      updatedAt: new Date(result._doc.updatedAt).toISOString()
-    };
-  },
-  cancelBooking: async args => {
-    try {
-      const booking = await Booking.findById(args.bookingId).populate('event');
-      const event = {
-        ...booking.event._doc,
-        _id: booking.event.id,
-        creator: user.bind(this, booking.event._doc.creator)
-      };
-      await Booking.deleteOne({ _id: args.bookingId });
-      return event;
-    } catch (err) {
-      throw err;
-    }
-  }
-};
+module.exports = rootResolver;
